Split view to separate file, restructure exports
