+++ examples/apps/spaces/src/app.ts
@@ -5,9 +5,12 @@
 
 import { getTinyliciousContainer } from "@fluid-experimental/get-container";
 import { getDefaultObjectFromContainer } from "@fluidframework/aqueduct";
+import React from "react";
+import ReactDOM from "react-dom";
 
 import { Spaces } from "./fluid-object";
 import { SpacesContainer } from "./container";
+import { AppView } from "./appView";
 
 // Re-export everything
 export { Spaces as SpacesExample, SpacesContainer };
@@ -34,7 +37,10 @@ async function start() {
     // For now we will just reach into the FluidObject to render it
     const contentDiv = document.getElementById("content");
     if (contentDiv !== null) {
-        defaultObject.render(contentDiv);
+        ReactDOM.render(
+            React.createElement(AppView, { model: defaultObject }),
+            contentDiv,
+        );
     }
 
     // Setting "fluidStarted" is just for our test automation
